+++ AshborneGame/_Core/Data/BOCS/ItemSystem/ItemCapabilities/ICanDamage.cs
@@ -3,6 +3,6 @@ namespace AshborneGame._Core.Data.BOCS.ItemSystem.ItemCapabilities
 {
     public interface ICanDamage
     {
-        int BaseDamage { get; set; }
+        double BaseDamage { get; set; }
     }
 }
