+++ AshborneGame/_Core/Player/StatHolder.cs
@@ -5,38 +5,38 @@ namespace AshborneGame._Core._Player
     public class StatHolder
     {
         public PlayerStatType Type { get; private set; }
-        public int BaseValue { get; private set; }
-        public int BonusValue { get; private set; } = 0;
-        public int Total => BaseValue + BonusValue;
+        public double BaseValue { get; private set; }
+        public double BonusValue { get; private set; } = 0;
+        public double Total => BaseValue + BonusValue;
 
-        public StatHolder(PlayerStatType statType, int initialValue = 0)
+        public StatHolder(PlayerStatType statType, double initialValue = 0)
         {
             Type = statType;
             BaseValue = initialValue;
         }
 
-        public void SetBase(int value)
+        public void SetBase(double value)
         {
             BaseValue = value;
         }
 
-        public void IncreaseBase(int value)
+        public void IncreaseBase(double value)
         {
             BaseValue += value;
         }
 
-        public void LowerBase(int value)
+        public void LowerBase(double value)
         {
             BaseValue -= value;
             if (BaseValue < 0) BaseValue = 0; // Ensure base value doesn't go negative
         }
 
-        public void AddBonus(int bonus)
+        public void AddBonus(double bonus)
         {
             BonusValue += bonus;
         }
 
-        public void RemoveBonus(int bonus)
+        public void RemoveBonus(double bonus)
         {
             BonusValue -= bonus;
             if (BonusValue < 0) BonusValue = 0; // Ensure bonus value doesn't go negative
